Moved responsemodel from model to logic module, removed unused dependencies, added used undeclared dependencies, moved version of dependencies to dependencyManagement block
