+++ steam-api-logic/src/main/java/com/steamext/steam/api/logic/client/SteamClientImpl.java
@@ -7,8 +7,8 @@ import com.steamext.steam.api.logic.exceptions.SteamHttpClientException;
 import com.steamext.steam.api.logic.client.http.request.GetRsaKeySteamHttpRequest;
 import com.steamext.steam.api.logic.client.http.request.LoginSteamHttpRequest;
 import com.steamext.steam.api.model.requestmodel.UserCredentials;
-import com.steamext.steam.api.model.requestmodel.UserLoginInfo;
-import com.steamext.steam.api.model.responsemodel.RsaDataContainer;
+import com.steamext.steam.api.logic.model.responsemodel.UserLoginInfo;
+import com.steamext.steam.api.logic.model.responsemodel.RsaDataContainer;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
