+++ steam-api-logic/src/main/java/com/steamext/steam/api/logic/client/http/request/LoginSteamHttpRequest.java
@@ -2,8 +2,8 @@ package com.steamext.steam.api.logic.client.http.request;
 
 import com.steamext.steam.api.logic.utils.SteamPasswordRSAEncode;
 import com.steamext.steam.api.model.requestmodel.UserCredentials;
-import com.steamext.steam.api.model.requestmodel.UserLoginInfo;
-import com.steamext.steam.api.model.responsemodel.RsaDataContainer;
+import com.steamext.steam.api.logic.model.responsemodel.UserLoginInfo;
+import com.steamext.steam.api.logic.model.responsemodel.RsaDataContainer;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
