+++ steam-api-model/pom.xml
@@ -15,13 +15,6 @@
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
         </dependency>
-
-        <!-- temporary -->
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>2.11.1</version>
-        </dependency>
     </dependencies>
 
     <artifactId>steam-api-model</artifactId>
