+++ steam-api-model/src/main/java/com/steamext/steam/api/model/requestmodel/UserPageInfo.java
@@ -2,8 +2,10 @@ package com.steamext.steam.api.model.requestmodel;
 
 import lombok.Data;
 
+import java.io.Serializable;
+
 @Data
-public class UserPageInfo {
+public class UserPageInfo implements Serializable {
     private String username;
     private String realname;
     private String address;
